early and late binding
binding is a process which convert the identifiers in addresses. binding is done for each variable and function which helps to matching the call with right right function defination. it is done either at run time or compile time

as a name early binding directly associate the address of the function to the function call. it replaces the call with a machine language instruction that tells mainframe to leap  to the address  of function.